+++ frontend/templates/frontend/index.html
@@ -267,11 +267,11 @@
                                     </td>
                                 </tr>
                                 <tr>
-                                    <td class="ddp-hover-tooltip">
-                                        <span class="ddp-txt-long ">localhost
+                                    <td className="ddp-hover-tooltip">
+                                        <span className="ddp-txt-long ">localhost
                                         </span>
-                                        <div class="ddp-ui-tooltip-info ddp-down">
-                                            <em class="ddp-icon-view-top"></em>
+                                        <div className="ddp-ui-tooltip-info ddp-down">
+                                            <em className="ddp-icon-view-top"></em>
                                             HOST : localhost <br/>
                                             PORT : 8090
                                         </div>
@@ -280,17 +280,17 @@
                                         Overlord : 8090
                                     </td>
                                     <td>
-                                        <span class="ddp-data-state ddp-success">SUCCESS</span>
+                                        <span className="ddp-data-state ddp-success">SUCCESS</span>
                                     </td>
                                     <td>
                                     </td>
                                 </tr>
                                 <tr>
-                                    <td class="ddp-hover-tooltip">
-                                        <span class="ddp-txt-long ">localhost
+                                    <td className="ddp-hover-tooltip">
+                                        <span className="ddp-txt-long ">localhost
                                         </span>
-                                        <div class="ddp-ui-tooltip-info ddp-down">
-                                            <em class="ddp-icon-view-top"></em>
+                                        <div className="ddp-ui-tooltip-info ddp-down">
+                                            <em className="ddp-icon-view-top"></em>
                                             HOST : localhost <br/>
                                             PORT : 8091
                                         </div>
@@ -299,17 +299,17 @@
                                         MiddleManager : 8091
                                     </td>
                                     <td>
-                                        <span class="ddp-data-state ddp-success">SUCCESS</span>
+                                        <span className="ddp-data-state ddp-success">SUCCESS</span>
                                     </td>
                                     <td>
                                     </td>
                                 </tr>
                                 <tr>
-                                    <td class="ddp-hover-tooltip">
-                                        <span class="ddp-txt-long ">localhost
+                                    <td className="ddp-hover-tooltip">
+                                        <span className="ddp-txt-long ">localhost
                                         </span>
-                                        <div class="ddp-ui-tooltip-info ddp-down">
-                                            <em class="ddp-icon-view-top"></em>
+                                        <div className="ddp-ui-tooltip-info ddp-down">
+                                            <em className="ddp-icon-view-top"></em>
                                             HOST : localhost <br/>
                                             PORT : 8100
                                         </div>
@@ -318,17 +318,17 @@
                                         MiddleManager : 8100
                                     </td>
                                     <td>
-                                        <span class="ddp-data-state ddp-success">SUCCESS</span>
+                                        <span className="ddp-data-state ddp-success">SUCCESS</span>
                                     </td>
                                     <td>
                                     </td>
                                 </tr>
                                 <tr>
-                                    <td class="ddp-hover-tooltip">
-                                        <span class="ddp-txt-long ">localhost
+                                    <td className="ddp-hover-tooltip">
+                                        <span className="ddp-txt-long ">localhost
                                         </span>
-                                        <div class="ddp-ui-tooltip-info ddp-down">
-                                            <em class="ddp-icon-view-top"></em>
+                                        <div className="ddp-ui-tooltip-info ddp-down">
+                                            <em className="ddp-icon-view-top"></em>
                                             HOST : localhost <br/>
                                             PORT : 8101
                                         </div>
@@ -337,7 +337,7 @@
                                         MiddleManager : 8101
                                     </td>
                                     <td>
-                                        <span class="ddp-data-state ddp-success">SUCCESS</span>
+                                        <span className="ddp-data-state ddp-success">SUCCESS</span>
                                     </td>
                                     <td>
                                     </td>
@@ -367,7 +367,16 @@
                         <button class="ddp-ui-edit-button"><a href="#top">Top</a></button>
                     </div>
 
+
                 </div>
+                <!--//nodes tab -->
+
+{#                <div id="nodes_coordinator" class="ddp-clear"></div>#}
+{#                <div id="nodes_broker" class="ddp-clear"></div>#}
+{#                <div id="nodes_historical" class="ddp-clear"></div>#}
+{#                <div id="nodes_overlord" class="ddp-clear"></div>#}
+{#                <div id="nodes_middleManager" class="ddp-clear"></div>#}
+                <!--//nodes tab -->
 
             </div>
