+++ init.sh
@@ -1,5 +1,5 @@
 
 npm run dev
 
-python manage.py runserver 0.0.0.0:8002
+python manage.py runserver 0.0.0.0:8001
